renamed component file to index.js
