+++ src/Components/CreateDogList/index.js
@@ -1,20 +1,20 @@
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import './CreateDogList.styles.scss';
-import { connect, useDispatch } from 'react-redux';
+import { connect, useDispatch, useSelector } from 'react-redux';
 import InfiniteScroll from 'react-infinite-scroll-component';
-import Loading from '../Loading/Loading';
-import DogCardTemplateList from '../DogCardTemplateList/DogCardTemplateList';
-import { fetchNameDataStart } from '../../Redux/Names/NamesRedux';
-import { getDogData } from '../../const/selectors/dog';
-import { changeLoading, updateOffset } from '../../Redux/Dog/DogRedux';
+import Loading from '../Loading';
+import DogCardTemplateList from '../DogCardTemplateList';
+import { fetchNameDataStart } from '../../Redux/names/index';
+import { getDogData } from '../../Redux/selectors/dog';
+import { changeLoading, updateOffset } from '../../Redux/dog/index';
 
 function CreateDogList({ isDogLoading }) {
   const dispatch = useDispatch();
   const [offsetAmount, setOffsetAmount] = useState(0);
   const dataContinues = true;
   const limitperPage = 20;
-  const completedDogData = getDogData();
+  const completedDogData = useSelector(getDogData());
 
   //* This function will execute when button is pressed
   //* it will activate the saga cycle
