+++ src/Components/NewProfile/index.js
@@ -5,8 +5,8 @@ import { useTranslation } from 'react-i18next';
 import './NewProfile.styles.scss';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faEnvelope, faPhone, faUser } from '@fortawesome/free-solid-svg-icons';
-import { updateUser } from '../../Redux/User/UserSlice';
-import { getAllUserInfo } from '../../const/selectors/user';
+import { updateUser } from '../../Redux/user/index';
+import { getAllUserInfo } from '../../Redux/selectors/user';
 
 function NewProfile() {
   const dispatch = useDispatch();
